Refactor arco (#117)
* move all ARCOERA5 code in one module

I renamed the model to a more explicict arcoera5 as ERA5 is the only provided
dataset. Consequently, the global variables used for validation and name
mapping were also renamed as they are only needed for ERA5, so this prefix can
be dropped. I made the variables private as I don't think this implementation
aspect should be exposed to the user, when they load this single module.

* remove a level of folder nesting, make ERA5 valid variables a list
* add tests for ARCOERA5 class, mark heavy ones as slow

The slow tests will load all variables, which will takes more time and much
more memory (~18GB).

* exclude slow tests from CI workflow
* add ARCOERA5 class to the documentation
* add arcoera5 constants to the API documentation