Add a data accessor for the WeatherBench2 datasets
The WeatherBench2 dataset is now used to ease creation of index classes for
each dataset of the WeatherBench2 project. I used __init_subclass__ to encode
the verification code to ensure derived classes have the right class attributes
to setup variables and levels validation in the constructor, and provide the
dataset URL.

ERA5 is provided as the first example.

Display in logs the storage used by downloaded dataset